+++ src/asserts/assert_named_export.ts
@@ -1,4 +1,5 @@
-import { ExtractedExport, NoExportError } from '@exercism/static-analysis'
+import type { ExtractedExport } from '@exercism/static-analysis'
+import { NoExportError } from '@exercism/static-analysis'
 import type { TSESTree } from '@typescript-eslint/typescript-estree'
 import { extractNamedExport } from '~src/extracts/extract_named_export'
 
@@ -19,7 +20,7 @@ export function assertNamedExport(
 ): ExtractedExport | never {
   // Find the export
   const fn = Array.isArray(root)
-    ? root.find((fn) => fn.exported === exported)
+    ? (root as readonly ExtractedExport[]).find((e) => e.exported === exported)
     : extractNamedExport(exported, root as Node)
 
   // Does it exist?