+++ src/output/processor/LogOutput.ts
@@ -1,6 +1,7 @@
 import { getProcessLogger } from '@exercism/static-analysis'
 import type { OutputProcessor } from '~src/interface'
 
+// eslint-disable-next-line @typescript-eslint/naming-convention
 export const LogOutput: OutputProcessor = async (
   previous: Promise<string>
 ): Promise<string> => {