+++ src/output/processor/PassThroughOutput.ts
@@ -1,5 +1,6 @@
-import { OutputProcessor } from '~src/interface'
+import type { OutputProcessor } from '~src/interface'
 
+// eslint-disable-next-line @typescript-eslint/naming-convention
 export const PassThroughOutput: OutputProcessor = async (
   previous: Promise<string>
 ): Promise<string> => {