+++ test/ref.d.ts
@@ -1,4 +1,0 @@
-interface MockedFs {
-  __setMockFiles(files: { [path: string]: string }): void
-  __getWrittenFiles(): { [dir: string]: { [file: string]: string } }
-}